+++ CHANGES.rst
@@ -10,7 +10,7 @@
 - Change `zope.interface.implements[Only]` and `zope.component.adapts` to
   class decorators.
 
-- Adapt the code to `icemac.addressbook >= 4.2`.
+- Adapt the code to `icemac.addressbook >= 5.0`.
 
 
 2.2 (2017-05-16)