+++ CHANGES.rst
@@ -7,6 +7,9 @@
 
 - Add breadcrumbs.
 
+- Fix rendering of day numbers in calendar for read-only users to be the same
+  size as for r/w users.
+
 - Change `zope.interface.implements[Only]` and `zope.component.adapts` to
   class decorators.